requirements.txt
asgiref==3.5.2
cycler==0.11.0
Django==4.0.6
django-analytical==3.1.0
django-debug-toolbar==3.6.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
fonttools==4.37.1
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pip==22.2.2
pyparsing==3.0.9
python-dateutil==2.8.2
setuptools==58.1.0
six==1.16.0
sqlparse==0.4.2
